Resolve Todos, tweak docs (#12553)
* wip

* change reserveAmounts to use IERC20

* only recalculate billingParams if payment tokens differ

* update readme

* fix migrate, regenerate wrappers

+++ contracts/src/v0.8/automation/dev/test/AutomationRegistry2_3.t.sol
@@ -187,7 +187,7 @@ contract Withdraw is SetUp {
     assertGt(linkToken.balanceOf(address(registry)), 0);
 
     //check the link available for payment is the link balance
-    assertEq(registry.linkAvailableForPayment(), linkToken.balanceOf(address(registry)));
+    assertEq(uint256(registry.linkAvailableForPayment()), linkToken.balanceOf(address(registry)));
   }
 
   function testWithdrawLinkFeesRevertsBecauseOnlyFinanceAdminAllowed() public {
